refactor: move file from apps to components
