+++ apps/front/src/components/table/select/index.tsx
@@ -1,0 +1,75 @@
+import { Table } from "@/components/ui/datat-table";
+import { ColumnDef, ColumnFiltersState, getCoreRowModel, getFilteredRowModel, getPaginationRowModel, getSortedRowModel, SortingState, useReactTable, VisibilityState } from "@tanstack/react-table";
+import React, { useState } from "react";
+import SelectTableFooter from "./footer";
+import SelectTableBody from "./body";
+import SelectTableHeader from "./header";
+import SelectTableFilters from "./filters";
+
+/**
+ * Props for the SelectTable component.
+ *
+ * @template T - The type of data in the table.
+ * @property {T[]} data - The array of data to be displayed in the table.
+ * @property {ColumnDef<T>[]} columns - The array of column definitions for the table.
+ * @property {string} [filterSearchPlaceholder] - The placeholder text for the filter search input.
+ * @property {keyof T} [filterSearchColumn] - The key of the column to be used for filtering the search.
+ * @property {boolean} [filterColumn] - Indicates whether filtering should be enabled for the table.
+ */
+export interface SelectTableProps<T> {
+  data: T[];
+  columns: ColumnDef<T>[];
+  filterSearchPlaceholder?: string;
+  filterSearchColumn?: keyof T;
+  filterColumn?: boolean;
+  actionComponent?: (props: { table: ReturnType<typeof useReactTable<T>> }) => React.ReactNode;
+}
+
+export default function SelectTable<T>(
+  { data, columns, filterSearchColumn, filterSearchPlaceholder, filterColumn, actionComponent }: SelectTableProps<T>
+) {
+  const [sorting, setSorting] = useState<SortingState>([])
+  const [columnFilters, setColumnFilters] = useState<ColumnFiltersState>([])
+  const [columnVisibility, setColumnVisibility] = useState<VisibilityState>({})
+  const [rowSelection, setRowSelection] = useState({})
+
+  const table = useReactTable({
+    data: data,
+    columns: columns,
+    onSortingChange: setSorting,
+    onColumnFiltersChange: setColumnFilters,
+    getCoreRowModel: getCoreRowModel(),
+    getPaginationRowModel: getPaginationRowModel(),
+    getSortedRowModel: getSortedRowModel(),
+    getFilteredRowModel: getFilteredRowModel(),
+    onColumnVisibilityChange: setColumnVisibility,
+    onRowSelectionChange: setRowSelection,
+    state: {
+      sorting,
+      columnFilters,
+      columnVisibility,
+      rowSelection,
+    },
+  })
+
+  return (
+    <div className="w-full">
+      <div className="flex flex-wrap items-center justify-center py-4 gap-3">
+        {actionComponent && actionComponent({ table })}
+        <SelectTableFilters
+          table={table}
+          filterSearchColumn={filterSearchColumn}
+          filterSearchPlaceholder={filterSearchPlaceholder}
+          filterColumns={filterColumn}
+        />
+      </div>
+      <div className="rounded-md border">
+        <Table>
+          <SelectTableHeader table={table} />
+          <SelectTableBody table={table} />
+        </Table>
+      </div>
+      <SelectTableFooter table={table} />
+    </div>
+  )
+}
